Switch away from data-default-class
data-default used to be split into several smaller packages, with
data-default itself just depending on them all and re-exporting the
class. These days, data-default contains everything and
data-default-class just exists for easier upgrades, it just depends on
data-default itself.

Note that the split stll carries a risk of confusion downstream; we
cannot help this. Suppose one package uses the new data-default 0.8 and
another uses an old data-default-class <0.2. The packages could both be
dependencies, or one is a (dependency of a) dependency and the other is
the primary package. You can get a bewildering error message
"No instance for (Default T)" because there exist two different classes
both called `Default` and the error message doesn't qualify the class
name.